refactor: 移动本体到文件夹中以使用__version__表示版本；最低版本根据rich，设置为3.7；使用GitHub提供的workflow
